Redes_Ruteo/requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
shapely==2.0.6
python-dotenv==1.0.1
geojson==3.1.0
pyproj==3.6.1
